refactor: split StressTest.tsx into focused sub-components (#379)
Extract 5 sub-components from StressTest.tsx (1318 → 619 lines):
- ShockSliders.tsx (166 lines) — per-asset shock % sliders
- PresetScenarioSelector.tsx (93 lines) — preset scenario button grid
- StressResultsTable.tsx (167 lines) — per-holding stress impact table
- ScenarioComparison.tsx (267 lines) — side-by-side scenario comparison panel
- ResilienceSummary.tsx (193 lines) — resilience score + summary metrics

No logic changes — pure structural refactor.
