+++ frontend/components/StressTest.tsx
@@ -11,39 +11,29 @@ import {
   Cell,
   ReferenceLine,
 } from 'recharts';
-import { HelpCircle } from 'lucide-react';
 import { usePortfolio } from '../hooks/usePortfolio';
 import { useStressTest } from '../hooks/useStressTest';
 import {
   createPresetScenarioInfo,
   createPresetScenarios,
   fxShockKey,
-  ASSET_TYPE_CONFIG,
   ACCOUNT_OPTIONS,
 } from '../lib/constants';
 import { formatCurrency, formatPercent, formatCompact } from '../lib/format';
 import { pnlColor } from '../lib/colors';
 import { EmptyState } from './ui/EmptyState';
 import { Select } from './ui/Select';
-import { CollapsiblePanel } from './ui/CollapsiblePanel';
 import { StressTestInfo } from './StressTestInfo';
 import { config } from '../lib/config';
-import type {
-  AccountType,
-  AssetType,
-  PortfolioSnapshot,
-  StressScenario,
-  StressScenarioInfo,
-} from '../types/portfolio';
+import { ShockSliders } from './ShockSliders';
+import { PresetScenarioSelector } from './PresetScenarioSelector';
+import { StressResultsTable } from './StressResultsTable';
+import { ScenarioComparison } from './ScenarioComparison';
+import { ResilienceSummary } from './ResilienceSummary';
+import type { AccountType, AssetType, PortfolioSnapshot, StressScenario } from '../types/portfolio';
 
 // ─── Shock state keyed as the scenario.shocks keys ───────────────────────────
 type ShockMap = Record<string, number>; // values are decimals e.g. -0.20
-
-const ASSET_SLIDERS: { key: string; label: string }[] = [
-  { key: 'stock', label: 'Stocks' },
-  { key: 'etf', label: 'ETFs' },
-  { key: 'crypto', label: 'Crypto' },
-];
 
 const ZERO_SHOCKS: ShockMap = {
   stock: 0,
@@ -80,483 +70,6 @@ const SECTION_TITLE: React.CSSProperties = {
   paddingBottom: 6,
   borderBottom: '1px solid var(--border-subtle)',
 };
-
-const TD: React.CSSProperties = {
-  padding: '6px 10px',
-  fontSize: 11,
-  borderBottom: '1px solid var(--border-subtle)',
-  borderRight: '1px solid var(--border-subtle)',
-  whiteSpace: 'nowrap',
-};
-
-// ─── Slider ───────────────────────────────────────────────────────────────────
-function ShockSlider({
-  label,
-  value,
-  onChange,
-}: {
-  label: string;
-  value: number;
-  onChange: (v: number) => void;
-}) {
-  const pct = Math.round(value * 100);
-  const color =
-    value > 0 ? 'var(--color-gain)' : value < 0 ? 'var(--color-loss)' : 'var(--color-accent)';
-  // Fill gradient position: 0 = -50%, 50 = 0%, 100 = +50%
-  const pos = ((value * 100 + 50) / 100) * 100; // 0-100
-  const gradStop =
-    value < 0
-      ? `var(--color-loss) 0%, var(--color-loss) ${pos}%, var(--border-primary) ${pos}%`
-      : value > 0
-        ? `var(--border-primary) 0%, var(--border-primary) ${pos}%, var(--color-gain) ${pos}%`
-        : `var(--border-primary) 0%`;
-
-  return (
-    <div style={{ marginBottom: 14 }}>
-      <div style={{ display: 'flex', justifyContent: 'space-between', marginBottom: 5 }}>
-        <span
-          style={{ fontSize: 12, color: 'var(--text-secondary)', fontFamily: 'var(--font-mono)' }}
-        >
-          {label}
-        </span>
-        <span
-          style={{
-            fontSize: 12,
-            fontFamily: 'var(--font-mono)',
-            fontWeight: 600,
-            color,
-            minWidth: 50,
-            textAlign: 'right',
-          }}
-        >
-          {pct >= 0 ? '+' : ''}
-          {pct}%
-        </span>
-      </div>
-      <input
-        type="range"
-        min={-50}
-        max={50}
-        step={1}
-        value={pct}
-        onChange={(e) => onChange(parseInt(e.target.value) / 100)}
-        style={{
-          width: '100%',
-          height: 4,
-          appearance: 'none',
-          WebkitAppearance: 'none',
-          background: `linear-gradient(to right, ${gradStop})`,
-          outline: 'none',
-          cursor: 'pointer',
-          borderRadius: 0,
-        }}
-      />
-    </div>
-  );
-}
-
-// ─── Compute stressed value for a scenario inline ─────────────────────────────
-function computeScenarioLocally(
-  snapshot: PortfolioSnapshot,
-  scenario: StressScenario
-): {
-  stressedValue: number;
-  totalImpact: number;
-  totalImpactPercent: number;
-  topImpacted: { symbol: string; impact: number }[];
-} {
-  let totalStressed = 0;
-  const breakdown: { symbol: string; impact: number }[] = [];
-
-  for (const h of snapshot.holdings) {
-    const assetShock = scenario.shocks[h.assetType] ?? 0;
-    const fxKey = fxShockKey(h.currency, snapshot.baseCurrency);
-    const fxShock =
-      h.currency.toUpperCase() === snapshot.baseCurrency.toUpperCase()
-        ? 0
-        : (scenario.shocks[fxKey] ?? 0);
-
-    const currentValue = h.marketValueCad;
-    const stressedValue = currentValue * (1 + assetShock) * (1 + fxShock);
-    const impact = stressedValue - currentValue;
-    totalStressed += stressedValue;
-    breakdown.push({ symbol: h.symbol, impact });
-  }
-
-  const currentValue = snapshot.totalValue;
-  const totalImpact = totalStressed - currentValue;
-  const totalImpactPercent = currentValue !== 0 ? (totalImpact / currentValue) * 100 : 0;
-
-  const topImpacted = [...breakdown]
-    .sort((a, b) => Math.abs(b.impact) - Math.abs(a.impact))
-    .slice(0, 2);
-
-  return { stressedValue: totalStressed, totalImpact, totalImpactPercent, topImpacted };
-}
-
-// ─── Scenario comparison table ────────────────────────────────────────────────
-function ScenarioComparison({
-  portfolio,
-  scenarios,
-}: {
-  portfolio: PortfolioSnapshot;
-  scenarios: StressScenarioInfo[];
-}) {
-  const baseCurrency = portfolio.baseCurrency;
-
-  const rows = useMemo(
-    () =>
-      scenarios.map((s) => {
-        const result = computeScenarioLocally(portfolio, s);
-        return {
-          name: s.name,
-          description: s.description,
-          stressedValue: result.stressedValue,
-          totalImpact: result.totalImpact,
-          totalImpactPercent: result.totalImpactPercent,
-          topImpacted: result.topImpacted,
-        };
-      }),
-    [portfolio, scenarios]
-  );
-
-  return (
-    <div style={{ ...PANEL, marginBottom: 1 }}>
-      <div style={SECTION_TITLE}>Scenario Comparison</div>
-      <div style={{ overflowX: 'auto' }}>
-        <table style={{ width: '100%', borderCollapse: 'collapse', fontSize: 11 }}>
-          <thead>
-            <tr>
-              {[
-                'Scenario',
-                `Stressed Value (${baseCurrency})`,
-                `Impact (${baseCurrency})`,
-                'Impact (%)',
-                'Top Impacted Holdings',
-              ].map((col) => (
-                <th
-                  key={col}
-                  style={{
-                    ...TD,
-                    background: 'var(--bg-surface-alt)',
-                    color: 'var(--text-secondary)',
-                    fontWeight: 600,
-                    textTransform: 'uppercase',
-                    letterSpacing: '0.06em',
-                    fontSize: 9,
-                    textAlign:
-                      col === 'Scenario' || col === 'Top Impacted Holdings' ? 'left' : 'right',
-                    borderBottom: '1px solid var(--border-primary)',
-                  }}
-                >
-                  {col}
-                </th>
-              ))}
-            </tr>
-          </thead>
-          <tbody>
-            {/* Current value reference row */}
-            <tr style={{ background: 'var(--bg-surface-alt)' }}>
-              <td
-                style={{
-                  ...TD,
-                  fontFamily: 'var(--font-mono)',
-                  fontWeight: 700,
-                  color: 'var(--text-primary)',
-                }}
-              >
-                Current Value
-              </td>
-              <td
-                style={{
-                  ...TD,
-                  textAlign: 'right',
-                  fontFamily: 'var(--font-mono)',
-                  fontWeight: 700,
-                  color: 'var(--text-primary)',
-                }}
-              >
-                {formatCurrency(portfolio.totalValue, baseCurrency)}
-              </td>
-              <td
-                style={{
-                  ...TD,
-                  textAlign: 'right',
-                  color: 'var(--text-muted)',
-                  fontFamily: 'var(--font-mono)',
-                }}
-              >
-                —
-              </td>
-              <td
-                style={{
-                  ...TD,
-                  textAlign: 'right',
-                  color: 'var(--text-muted)',
-                  fontFamily: 'var(--font-mono)',
-                }}
-              >
-                —
-              </td>
-              <td style={{ ...TD, color: 'var(--text-muted)', borderRight: 'none' }}>—</td>
-            </tr>
-
-            {/* Scenario rows */}
-            {rows.map((row, i) => (
-              <tr
-                key={row.name}
-                style={{ background: i % 2 === 0 ? 'var(--bg-surface)' : 'var(--bg-surface-alt)' }}
-              >
-                <td style={{ ...TD, fontFamily: 'var(--font-mono)' }}>
-                  <div style={{ fontWeight: 600, color: 'var(--text-primary)' }}>{row.name}</div>
-                  <div
-                    style={{
-                      fontSize: 10,
-                      color: 'var(--text-muted)',
-                      marginTop: 2,
-                      fontFamily: 'var(--font-sans)',
-                      whiteSpace: 'normal',
-                      maxWidth: 220,
-                    }}
-                  >
-                    {row.description}
-                  </div>
-                </td>
-                <td
-                  style={{
-                    ...TD,
-                    textAlign: 'right',
-                    fontFamily: 'var(--font-mono)',
-                    color: pnlColor(row.totalImpact),
-                  }}
-                >
-                  {formatCurrency(row.stressedValue, baseCurrency)}
-                </td>
-                <td
-                  style={{
-                    ...TD,
-                    textAlign: 'right',
-                    fontFamily: 'var(--font-mono)',
-                    fontWeight: 600,
-                    color: pnlColor(row.totalImpact),
-                  }}
-                >
-                  {row.totalImpact >= 0 ? '+' : ''}
-                  {formatCurrency(row.totalImpact, baseCurrency)}
-                </td>
-                <td
-                  style={{
-                    ...TD,
-                    textAlign: 'right',
-                    fontFamily: 'var(--font-mono)',
-                    fontWeight: 600,
-                    color: pnlColor(row.totalImpact),
-                  }}
-                >
-                  {formatPercent(row.totalImpactPercent)}
-                </td>
-                <td style={{ ...TD, borderRight: 'none' }}>
-                  {row.topImpacted.length === 0 ? (
-                    <span style={{ color: 'var(--text-muted)', fontFamily: 'var(--font-mono)' }}>
-                      —
-                    </span>
-                  ) : (
-                    <div style={{ display: 'flex', flexDirection: 'column', gap: 2 }}>
-                      {row.topImpacted.map((h) => (
-                        <div
-                          key={h.symbol}
-                          style={{
-                            fontFamily: 'var(--font-mono)',
-                            fontSize: 10,
-                            color: pnlColor(h.impact),
-                          }}
-                        >
-                          <span style={{ fontWeight: 700 }}>{h.symbol}</span>{' '}
-                          <span>
-                            {h.impact >= 0 ? '+' : ''}
-                            {formatCompact(h.impact)}
-                          </span>
-                        </div>
-                      ))}
-                    </div>
-                  )}
-                </td>
-              </tr>
-            ))}
-          </tbody>
-        </table>
-      </div>
-    </div>
-  );
-}
-
-// ─── Resilience summary ───────────────────────────────────────────────────────
-function ResilienceSummary({ portfolio }: { portfolio: PortfolioSnapshot | null }) {
-  if (!portfolio || portfolio.holdings.length === 0) return null;
-
-  const baseCurrency = portfolio.baseCurrency;
-
-  // Largest single-holding risk by weight
-  const largestHolding = [...portfolio.holdings].sort((a, b) => b.weight - a.weight)[0];
-
-  // Max hit: holding with highest market value (worst case if it goes to zero)
-  const maxHitHolding = [...portfolio.holdings].sort(
-    (a, b) => b.marketValueCad - a.marketValueCad
-  )[0];
-
-  // Diversification score: 1 - HHI, normalized
-  // Only non-cash holdings for diversification computation
-  const nonCash = portfolio.holdings.filter((h) => h.assetType !== 'cash');
-  const n = nonCash.length;
-  let diversificationScore = 0;
-  if (n > 1) {
-    const totalNonCash = nonCash.reduce((sum, h) => sum + h.marketValueCad, 0);
-    if (totalNonCash > 0) {
-      const hhi = nonCash.reduce((sum, h) => {
-        const w = h.marketValueCad / totalNonCash;
-        return sum + w * w;
-      }, 0);
-      diversificationScore = ((1 - hhi) / (1 - 1 / n)) * 100;
-    }
-  } else if (n === 1) {
-    diversificationScore = 0;
-  }
-
-  // FX exposure: % of portfolio in non-base-currency holdings
-  const fxExposureValue = portfolio.holdings
-    .filter((h) => h.currency.toUpperCase() !== baseCurrency.toUpperCase())
-    .reduce((sum, h) => sum + h.marketValueCad, 0);
-  const fxExposurePct =
-    portfolio.totalValue > 0 ? (fxExposureValue / portfolio.totalValue) * 100 : 0;
-
-  // Cash buffer: % of portfolio in cash positions
-  const cashValue = portfolio.holdings
-    .filter((h) => h.assetType === 'cash')
-    .reduce((sum, h) => sum + h.marketValueCad, 0);
-  const cashPct = portfolio.totalValue > 0 ? (cashValue / portfolio.totalValue) * 100 : 0;
-
-  const STAT_CARD: React.CSSProperties = {
-    background: 'var(--bg-surface)',
-    border: '1px solid var(--border-primary)',
-    padding: '14px 16px',
-    flex: '1 1 0',
-    minWidth: 160,
-  };
-
-  const STAT_LABEL: React.CSSProperties = {
-    fontSize: 10,
-    color: 'var(--text-muted)',
-    fontFamily: 'var(--font-mono)',
-    textTransform: 'uppercase',
-    letterSpacing: '0.08em',
-    marginBottom: 6,
-  };
-
-  const STAT_VALUE: React.CSSProperties = {
-    fontSize: 20,
-    fontFamily: 'var(--font-mono)',
-    fontWeight: 700,
-    color: 'var(--text-primary)',
-    lineHeight: 1.2,
-  };
-
-  const STAT_SUB: React.CSSProperties = {
-    fontSize: 11,
-    fontFamily: 'var(--font-mono)',
-    color: 'var(--text-secondary)',
-    marginTop: 4,
-  };
-
-  return (
-    <div style={{ marginTop: 1 }}>
-      <div
-        style={{
-          ...PANEL,
-          paddingBottom: 16,
-          background: 'var(--bg-surface)',
-        }}
-      >
-        <div style={SECTION_TITLE}>Portfolio Resilience</div>
-        <div
-          style={{ display: 'flex', gap: 1, flexWrap: 'wrap', background: 'var(--border-primary)' }}
-        >
-          {/* Largest single-holding risk */}
-          <div style={STAT_CARD}>
-            <div style={STAT_LABEL}>Largest Position</div>
-            <div style={STAT_VALUE}>
-              {largestHolding ? `${(largestHolding.weight * 100).toFixed(1)}%` : '—'}
-            </div>
-            <div style={STAT_SUB}>{largestHolding ? largestHolding.symbol : '—'} of portfolio</div>
-          </div>
-
-          {/* Max hit from one holding */}
-          <div style={STAT_CARD}>
-            <div style={STAT_LABEL}>Max Single-Holding Loss</div>
-            <div style={{ ...STAT_VALUE, color: 'var(--color-loss)' }}>
-              {maxHitHolding ? formatCompact(maxHitHolding.marketValueCad) : '—'}
-            </div>
-            <div style={STAT_SUB}>{maxHitHolding ? `${maxHitHolding.symbol} at zero` : '—'}</div>
-          </div>
-
-          {/* Diversification score */}
-          <div style={STAT_CARD}>
-            <div style={STAT_LABEL}>Diversification Score</div>
-            <div
-              style={{
-                ...STAT_VALUE,
-                color:
-                  diversificationScore >= 70
-                    ? 'var(--color-gain)'
-                    : diversificationScore >= 40
-                      ? 'var(--color-warning)'
-                      : 'var(--color-loss)',
-              }}
-            >
-              {n > 0 ? `${diversificationScore.toFixed(0)} / 100` : '—'}
-            </div>
-            <div style={STAT_SUB}>
-              {n > 0
-                ? diversificationScore >= 70
-                  ? 'Well diversified'
-                  : diversificationScore >= 40
-                    ? 'Moderate concentration'
-                    : 'Highly concentrated'
-                : 'No non-cash holdings'}
-            </div>
-          </div>
-
-          {/* FX exposure */}
-          <div style={STAT_CARD}>
-            <div style={STAT_LABEL}>Foreign Currency Exposure</div>
-            <div style={STAT_VALUE}>{fxExposurePct.toFixed(1)}%</div>
-            <div style={STAT_SUB}>
-              {formatCompact(fxExposureValue)} in non-{baseCurrency}
-            </div>
-          </div>
-
-          {/* Cash buffer */}
-          <div style={STAT_CARD}>
-            <div style={STAT_LABEL}>Cash Buffer</div>
-            <div
-              style={{
-                ...STAT_VALUE,
-                color:
-                  cashPct >= 5
-                    ? 'var(--color-gain)'
-                    : cashPct > 0
-                      ? 'var(--color-warning)'
-                      : 'var(--text-secondary)',
-              }}
-            >
-              {cashPct.toFixed(1)}%
-            </div>
-            <div style={STAT_SUB}>{formatCompact(cashValue)} in cash positions</div>
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-}
 
 // ─── Filter helper ────────────────────────────────────────────────────────────
 function applyHoldingFilters(
@@ -824,104 +337,20 @@ export function StressTest() {
             background: 'var(--bg-surface)',
           }}
         >
-          {/* Preset selector */}
-          <div style={{ marginBottom: 20 }}>
-            <div
-              style={{
-                display: 'flex',
-                alignItems: 'center',
-                justifyContent: 'space-between',
-                gap: 12,
-                marginBottom: 10,
-              }}
-            >
-              <div style={{ ...SECTION_TITLE, flex: 1, marginBottom: 0 }}>Preset Scenario</div>
-              {activePresetInfo && (
-                <button
-                  onClick={() => setInfoOpen(true)}
-                  style={{
-                    display: 'inline-flex',
-                    alignItems: 'center',
-                    gap: 6,
-                    background: 'transparent',
-                    border: '1px solid var(--border-primary)',
-                    color: 'var(--text-secondary)',
-                    padding: '5px 10px',
-                    cursor: 'pointer',
-                    fontFamily: 'var(--font-mono)',
-                    fontSize: 10,
-                    textTransform: 'uppercase',
-                    letterSpacing: '0.06em',
-                  }}
-                >
-                  <HelpCircle size={12} />
-                  Info
-                </button>
-              )}
-            </div>
-            <Select
-              value={presetName}
-              onChange={handlePresetChange}
-              options={presetNames.map((n) => ({ value: n, label: n }))}
-            />
-            {/* Scenario description */}
-            {presetName !== 'Custom' &&
-              (() => {
-                const preset = presetScenarioInfo.find((s) => s.name === presetName);
-                return preset ? (
-                  <div
-                    style={{
-                      marginTop: 8,
-                      fontSize: 10,
-                      color: 'var(--text-muted)',
-                      fontFamily: 'var(--font-sans)',
-                      lineHeight: 1.5,
-                    }}
-                  >
-                    {preset.description}
-                  </div>
-                ) : null;
-              })()}
-          </div>
+          <PresetScenarioSelector
+            presetName={presetName}
+            presetNames={presetNames}
+            scenarioInfo={presetScenarioInfo}
+            onSelect={handlePresetChange}
+            onInfoOpen={() => setInfoOpen(true)}
+          />
 
-          {/* Asset class shocks */}
-          <div style={{ marginBottom: 20 }}>
-            <div style={SECTION_TITLE}>Asset Class Shocks</div>
-            {ASSET_SLIDERS.map(({ key, label }) => (
-              <ShockSlider
-                key={key}
-                label={label}
-                value={shocks[key] ?? 0}
-                onChange={(v) => handleSliderChange(key, v)}
-              />
-            ))}
-          </div>
-
-          {/* FX shocks */}
-          {activeFxSliders.length > 0 && (
-            <div>
-              <div style={SECTION_TITLE}>Currency Shocks</div>
-              <div
-                style={{
-                  fontSize: 10,
-                  color: 'var(--text-muted)',
-                  fontFamily: 'var(--font-mono)',
-                  marginBottom: 10,
-                }}
-              >
-                Positive = {baseCurrency} weakens. Foreign holdings convert into more {baseCurrency}
-                . Negative = {baseCurrency} strengthens.
-              </div>
-              {activeFxSliders.map(({ key, label }) => (
-                <ShockSlider
-                  key={key}
-                  label={label}
-                  value={shocks[key] ?? 0}
-                  onChange={(v) => handleSliderChange(key, v)}
-                />
-              ))}
-            </div>
-          )}
+          <ShockSliders
+            shocks={shocks}
+            onChange={handleSliderChange}
+            activeFxSliders={activeFxSliders}
+            baseCurrency={baseCurrency}
+          />
 
           {loading && (
             <div
@@ -1164,127 +593,11 @@ export function StressTest() {
 
           {/* Breakdown table */}
           {result && !allZero && (
-            <CollapsiblePanel title="Breakdown" defaultExpanded={true}>
-              <div style={{ overflow: 'auto', maxHeight: 360 }}>
-                <table style={{ width: '100%', borderCollapse: 'collapse', fontSize: 11 }}>
-                  <thead style={{ position: 'sticky', top: 0 }}>
-                    <tr>
-                      {[
-                        'Symbol',
-                        'Type',
-                        `Current Value (${baseCurrency})`,
-                        'Shock',
-                        `Stressed Value (${baseCurrency})`,
-                        `Impact (${baseCurrency})`,
-                        'Impact (%)',
-                      ].map((col) => (
-                        <th
-                          key={col}
-                          style={{
-                            ...TD,
-                            background: 'var(--bg-surface-alt)',
-                            color: 'var(--text-secondary)',
-                            fontWeight: 600,
-                            textTransform: 'uppercase',
-                            letterSpacing: '0.06em',
-                            fontSize: 9,
-                            textAlign: col === 'Symbol' || col === 'Type' ? 'left' : 'right',
-                            borderBottom: '1px solid var(--border-primary)',
-                          }}
-                        >
-                          {col}
-                        </th>
-                      ))}
-                    </tr>
-                  </thead>
-                  <tbody>
-                    {[...result.holdingBreakdown]
-                      .sort((a, b) => a.impact - b.impact)
-                      .map((h, i) => {
-                        const holding = filteredPortfolio?.holdings.find(
-                          (p) => p.id === h.holdingId
-                        );
-                        const bg = i % 2 === 0 ? 'var(--bg-surface)' : 'var(--bg-surface-alt)';
-                        return (
-                          <tr key={h.holdingId} style={{ background: bg }}>
-                            <td
-                              style={{
-                                ...TD,
-                                fontFamily: 'var(--font-mono)',
-                                fontWeight: 700,
-                                color: 'var(--text-primary)',
-                              }}
-                            >
-                              {h.symbol}
-                            </td>
-                            <td style={{ ...TD, color: 'var(--text-secondary)' }}>
-                              {holding
-                                ? (ASSET_TYPE_CONFIG[holding.assetType]?.label ?? holding.assetType)
-                                : '—'}
-                            </td>
-                            <td
-                              style={{
-                                ...TD,
-                                textAlign: 'right',
-                                fontFamily: 'var(--font-mono)',
-                                color: 'var(--text-secondary)',
-                              }}
-                            >
-                              {formatCurrency(h.currentValue, baseCurrency)}
-                            </td>
-                            <td
-                              style={{
-                                ...TD,
-                                textAlign: 'right',
-                                fontFamily: 'var(--font-mono)',
-                                color: pnlColor(h.shockApplied),
-                              }}
-                            >
-                              {h.shockApplied !== 0 ? formatPercent(h.shockApplied * 100) : '—'}
-                            </td>
-                            <td
-                              style={{
-                                ...TD,
-                                textAlign: 'right',
-                                fontFamily: 'var(--font-mono)',
-                                color: pnlColor(h.impact),
-                              }}
-                            >
-                              {formatCurrency(h.stressedValue, baseCurrency)}
-                            </td>
-                            <td
-                              style={{
-                                ...TD,
-                                textAlign: 'right',
-                                fontFamily: 'var(--font-mono)',
-                                fontWeight: 600,
-                                color: pnlColor(h.impact),
-                              }}
-                            >
-                              {h.impact !== 0
-                                ? `${h.impact >= 0 ? '+' : ''}${formatCurrency(h.impact, baseCurrency)}`
-                                : '—'}
-                            </td>
-                            <td
-                              style={{
-                                ...TD,
-                                textAlign: 'right',
-                                fontFamily: 'var(--font-mono)',
-                                color: pnlColor(h.impact),
-                                borderRight: 'none',
-                              }}
-                            >
-                              {h.impact !== 0
-                                ? formatPercent((h.impact / h.currentValue) * 100)
-                                : '—'}
-                            </td>
-                          </tr>
-                        );
-                      })}
-                  </tbody>
-                </table>
-              </div>
-            </CollapsiblePanel>
+            <StressResultsTable
+              result={result}
+              holdings={filteredPortfolio?.holdings ?? []}
+              baseCurrency={baseCurrency}
+            />
           )}
         </div>
       </div>
@@ -1295,27 +608,6 @@ export function StressTest() {
         scenario={activePresetInfo}
         onClose={() => setInfoOpen(false)}
       />
-
-      {/* Slider thumb global style injection */}
-      <style>{`
-        input[type='range']::-webkit-slider-thumb {
-          -webkit-appearance: none;
-          appearance: none;
-          width: 12px;
-          height: 12px;
-          background: var(--color-accent);
-          border-radius: 0;
-          cursor: pointer;
-        }
-        input[type='range']::-moz-range-thumb {
-          width: 12px;
-          height: 12px;
-          background: var(--color-accent);
-          border-radius: 0;
-          border: none;
-          cursor: pointer;
-        }
-      `}</style>
     </div>
   );
 }
